requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
pandas>=2.1
openpyxl>=3.1
python-dateutil>=2.8
